Added support for roundboxes. The file extreme_round_768x576.xml is an example of round box usage. blue2_768x576.xml also looks a little bit different now, all other skins should still look like before.
You can now add <mask>-tags. All these masks and the selection are drawn
on the same layer. With that it is possible to undo the alpha changes in
a selection. So the alpha values in colors are changes to the background
image, not the current background (image+alpha mask)


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@1167 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5