Some stuff to work with the changes in datatypes and movie_config.
Also there is a seek to position code. Since the -ss flag from mplayer
is way to bad (seek to position 20:00 jumps to 35:xxx), I implemented
a small work around. Freevo moves forward in the video until we are
ahead for the start_position. Than we go backwards until it fits. This
is a bad hack (TM) and should be removed when seek to position works in
mplayer.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@1296 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5