Initial implementation of lirc support. The main motivation for this was to make it possible to run Freevo using the lirc client library as opposed to the direct socket reads being done now. Of course, the socket reads are better, faster and probably safer to use for now, so ignore this unless you're interested.
It behaves somewhat differently than the current implementation because
it waits for the remote button to be pressed instead of looping.

Finally, this might not be the best way to do it, it might be worthwhile
to do this in rc.py and skip the UDP part, which is useful for other
reasons.

Pylirc module is available at pylirc.sf.net
You need to put the lircrc file you make into /etc/freevo for now.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@1562 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5