Ugly hack to work around missing stop times. This is a slight improvement over my patch from yesterday, because it "guesses" the stop time of the shows.
Logic:

Since the only shows missing stop dates are those which appear to end at
midnight, we insert a stop date of midnight for those shows missing a stop
date.

This is not perfect. If you have a better solution, I'll be the first one to
put it in. Hopefully, this'll be fixed in XMLTV at some point and this hack
won't be necessary.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@1724 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5