The "informational" toolbar from my screenshot from yesterday. What this does is show a clock, and a mail icon indicating whether mail is available or not.
How it works:
    o We have a class, the horribly named "IdleTool"
    o When you initialize the class, it sets a counter.
    o You call IdleTool.poll in the idle loop in main.py
    o When the counter reaches a certain amount (in this case, 30 seconds or 300 steps of
        0.1 seconds) it updates the display.
    o Only two lines would need to be added to main.py to get it to work.

You need to add:

m = idle.IdleTool() before the 'while 1:' loop
and
m.poll() inside the loop

If new mail is found, the 'mail' icon undims, and when it's not there, it
disappears. I used the Python mailbox class, so while I'm only supporting
mbox here, it should be easy to add different ones.

All config information is in the __init__ and you need to change it to monitor
another box.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@1789 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5