Show the episode title in the description of a TV show if it's available.
Example:

---
The Simpsons
"The Parent Rap"   <-- the sub-title
Bart goes before a tough judge for stealing a police car.
---

Shows without a title (called a sub-title in xmltv) are shown as they were
before. Maybe we should use this when naming the Recorded files? It would
be an inexpensive way of making sure we don't re-record the same episodes.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@1936 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5