Merged Kyle Weston's Schedule Viewer/Editor patch.
Please note, it currently only works with recording files ending in '.avi'
This is temporary, but as a fix, there is a commented line on line 171,
if you use mpgs (like I do) you can enable line 171 and comment out line 170.

Since avi is the default, I'm leaving it that way, but I will use mpg myself.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@2063 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5