Changed the "Schedule Editor" to show up in the TV Submenu, along with "Guide" and "Recorded Shows" which makes a lot more sense then where it was before, which was also exceptionally well hidden.
To do this properly, I also had to move record_schedule into a class, subclassing
from Item, and so problems may and possibly will arise. I've tested it a little,
but please bang on this, because while it's a relatively minor change, it does
get things working inside the properly model, at least for a start.

Bug reports are expected and welcome :)


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@2601 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5