Ok, I think this is a pretty good preliminary version, and since it doesn't touch anything else, I think it could go into helpers.
Current issue:

The big one is that it tries to reimport entries even if they already exist,
it'll probably fail when that happens.

As it stands though, I imported 400 entries of varying tag quality with
md5 checksums into a new database in around 27s. I'm using the config entries
to figure out which folders to add and where to put the database.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@2865 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5