Added automatic XMLTV channel list code; the idea is that you can avoid editing the TV_CHANNELS thing yourself, if your xmltv grabber outputs properly formatted channel listings.
If this doesn't work, you can always fall back to adding it manually.

To use it, you need this in your local_conf.py
---
from util import getXMLTVChannels
TV_CHANNELS = getXMLTVChannels(XMLTV_FILE)
---

instead of

TV_CHANNELS = [('69 COMEDY', 'COMEDY', '69'),
               ...
              ]


I don't know if this will work for everyone, but it's working nicely for me,
and I know it's a constant setup difficulty, so please test it. If you don't
want to use it, don't do anything, it will not be used unless you add it
to your config.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@3528 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5