Changes to make the list update when you remove a scheduled recording.
Also made the buttons in the program display on top of one another instead
of side by side.  I will be adding more buttons (add to favorites, search
for more) and they will look and fit better this way.  I think they would
all be better together in a ListBox.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@3978 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5