Bugfix for missing 'menu' button; it was here before, and it was removed for some reason.
Most DVDs have multiple menus, but not all DVDs have a TitleMenu, so under
the previous event structure, you could only move to the topmost menu, and
only if it was designated as such.

This patch allows the old behaviour (from when I first implemented code to
support Dischi's Xine patches) which works with more DVDs.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@4126 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5