Added initial debconf support to the package.
What does this mean?

When you install the package, it will ask you questions about your display,
resolution, TV norm, and channel list and automatically generate a freevo.conf
file.

Currently, it will not touch the file if freevo.conf exists. In the future,
it should be able to figure out whether freevo.conf was made with debconf or
not and decide based on that.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@4140 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5