Fix a crash...
When you are going through a bunch of images manually (up/down) when you reach
the last image, it used to crash because the 'osd' object lacks a "drawstringframed"
method. By sending the proper initialized object (self.osd) it works nicely instead
showing the message over the image in the top right corner.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@4615 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5