Ok,  now it shows the "Coming Up" list anywhere in the TV menu. I think it fits, though it looks fairly ugly right now. I'm going to make it more flexible after I get some listings for 'tomorrow' since mine expire tonight.
I like this as a feature, but I'm wondering if someone has an idea on a
cleaner way to implement this. This is a little hackish, since "Coming Up"
isn't really a item "property" so it doesn't exactly fit in the object
model.

I did remove it from directory.py, so that's at least more logical.

Maybe we should have a general function in item.py to call extra
functions, or a way to embed python in the skin which isn't so nice.

Ideally, we need a different way to have "default" information in an info
area, as opposed to putting it in the item.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@4852 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5