Fixed to use proper audio context buttons.
Note: Is is possible to do this without adding code to playlist.py? All I need
is the filename and an event, but because no existing event sends the filename,
and the plugin doesn't know about it, I have to pass it along somehow.

Anyone want to chime in with an idea on how?


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@4960 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5