Handle filenames internally as unicode objects.
This does *NOT* affect filenames that have only ASCII chars, since the translation ASCII -> Unicode is painless. However this *DOES* affect files with accents, like �� (e acute, \xe9) and others.

I tested with Video, Images and Music modules, but *NOT* with Games, so if you have the games modules, give it a try.

It determines the encoding based on (in order) FREEVO_LOCALE, LANG and LC_ALL, which may have the form: "LANGUAGE_CODE.ENCODING", like "pt_BR.UTF-8", and others.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@5152 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5