This file will contain device keymaps (scancode : keycode) for input event devices.  These maps are the same ones that can be retrieved and set using the input tools found at: http://dl.bytesex.org/cvs-snapshots/input-20040421-115547.tar.gz
I used input-kbd to create the map for the hauppauge grey remote with the
pvr-250 and 350 cards.  I am unsure if the map is the same for other cards
that use this remote (DVB cards, nexus and nova for example).  Time will tell.

I expect users to send in their keymaps so we can provide nice defaults.

We will also support user defined keymaps for the device, I just need to
impliment that.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@6372 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5