New ivtv record plugin.
This plugin requires the "encoder" utility that comes with the CK series of
ivtv drivers found at http://67.18.1.101/~ckennedy/ivtv/.  You neeed at least
version 0.2.0-rc2s from the stable branch or 0.3.1o from the unstable branch
because we rely on a paramerter for passing the channel frequency that was
introduced in those versions.

To use:

plugin.activate('record.ivtv') # it defaults to ivtv0
or:
plugin.activate('record.ivtv', args='ivtv1')

Right now it only supports one card but I will be sending a patch for encoder
to support setting the device.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@6646 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5