Clean import for gui system. There are no gui imports in the __init__ anymore because this caused recursive imports: gui imports e.g. windows and in windows gui is imported again to get to the widgets.
So if a gui module like windows or widgets is needed, the correct
submodule needs to be imported. This is also needed for simple modules
like theme (renamed from theme_engine) and imagelib.

The only variables still direct in gui are width, weight and display, I
don't know were else to put them.


git-svn-id: svn://svn.freevo.org/freevo/trunk/freevo@7263 cc3e1ea1-1e01-0410-8d68-8b121e83a9d5