Make each application a kaa.candy.Layer
Each application now has its own layer. Additionally there are
layers for the background and OSD levels. How many layers besides
the applications exists id part of the XML definition. A layer may
not be scaled anymore but may have a scaled group inside it. This
is needed for the video player with an unscaled video but the OSD
defined for a specific geometry. Keeping an application is one
layer makes it possible to toggle applications and (in the future)
show a nice overview about running applications.