Fix osd hide/show code
The latest kaa.candy changes boke the hiding event because the
visible property was read by kaa.candy to hide the widget. Now
is should be more or less correct -- it won't be if the hiding
event takes longer than the showing.

Furthermore, application events are passed to the GUI.