Add basic 3D support for video overlay
If the video is in stereo mode (and marked as such) the OSD on top
of the video player will adjust to it and duplicated using the
clone widget. This only works for the normal OSD on top of the video
and not for other candy-based widgets such as the generic OSD for
messages.