Remove item.files
The files attribute based on Files was much too complex for a
simple tasks such as delete an item. The file_ops plugin still
needs some work, but it does not need the complex Files(), it
can gather everything from the item itself.

Note: files.py was already deleted in the prev. commit.