Fix race condition when creating a new menu
This commit should only add a menu refresh when the num_items
were calculated for the GUI. But together with an empty beacon
database and autoselect items the parent directory recreated
the item creating the menu. This results in a corrupt stack.
Two race conditions for this case are fixed now.